[REFACTOR] open modal
