+++ src/components/EnvManagement/EnvSection/EnvSection.tsx
@@ -1,16 +1,24 @@
-import { useState } from 'react';
 import * as S from './style';
 import { PublicTab } from '../..';
-import { ListWrapper, PaginationtWrapper, SectionWrapper } from '../../../style/Section';
+import { Modal } from '@mui/material';
 import StyledPagination from '../../Public/PaginationButton/PaginationButton';
-import { IEnvironment } from '../../../types/environment.types';
+import { ListWrapper, PaginationtWrapper, SectionWrapper } from '../../../style/Section';
 import EnvironmentRow from '../EnvironmentList/EnvironmentRow';
 import EnvironmentHeader from '../EnvironmentList/EnvironmentHeader';
-import { EPlatform } from '../../../lib/enum/platform';
 import useEnvironment from '../../../hooks/useEnvironment';
+import EnvironmentModal from '../../Modal/EnvironmentModal/EnvironmentModal';
+import DeleteModal from '../../Modal/DeleteModal/DeleteModal';
 
 const EnvSection = () => {
-  const { pageCount, environments, onClickCreateEnvironment } = useEnvironment();
+  const {
+    pageCount,
+    environments,
+    isOpenModifyModal,
+    isOpenDeleteModal,
+    toggleIsOpenModifyModal,
+    toggleIsOpenDeleteModal,
+    onClickCreateEnvironment,
+  } = useEnvironment();
 
   return (
     <SectionWrapper>
@@ -21,10 +29,23 @@ const EnvSection = () => {
       <ListWrapper>
         <EnvironmentHeader />
         <hr />
+
         {environments.map(environment => (
           <div key={environment.id}>
-            <EnvironmentRow environment={environment} />
+            <EnvironmentRow
+              environment={environment}
+              toggleIsOpenModifyModal={toggleIsOpenModifyModal}
+              toggleIsOpenDeleteModal={toggleIsOpenDeleteModal}
+            />
             <hr />
+
+            <Modal open={isOpenModifyModal} onClose={toggleIsOpenModifyModal}>
+              <EnvironmentModal type='modify' environmentValue={environment} />
+            </Modal>
+
+            <Modal open={isOpenDeleteModal} onClose={toggleIsOpenDeleteModal}>
+              <DeleteModal></DeleteModal>
+            </Modal>
           </div>
         ))}
       </ListWrapper>
